Cleaned up installer script; add verbose & dryrun option

+++ install.py
@@ -1,7 +1,20 @@
 #!/usr/bin/env python
 
 # install dotfiles by placing symlinks from ~/ to this git repo
-import os, shutil
+import os, shutil, sys
+
+flags = sys.argv[1:]
+
+if "--help" in flags:
+    print "Usage: install.py [flags]\nwhere the valid flags are:"
+    print "\t--help   \tDisplay this message and quit"
+    print "\t--verbose\tPrint all file operations"
+    print "\t--dryrun \tDon't perform file operations (implies --verbose)"
+    sys.exit(0)
+# Print out file operations, instead of performing them.
+dryrun = "--dryrun" in flags
+# --dryrun implies --verbose
+verbose = dryrun or "--verbose" in flags
 
 dotfiles = [".emacs",".cshrc.common",".dir_colors",".Xdefaults",".vimrc"]
 
@@ -22,5 +35,15 @@ for file in dotfiles:
             repl = os.path.join(home,file+".BAK."+str(i))
             cond = os.path.exists(repl)
             i += 1
-        shutil.move(target,repl)
-    os.symlink(source,target)
+        if verbose:
+            print "Moving %s to %s" % (target,repl)
+        if not dryrun:
+            shutil.move(target,repl)
+    if verbose:
+        print "Creating symlink from %s to %s\n--" % (source,target)
+    if not dryrun:
+        os.symlink(source,target)
+print "Dotfile symlinks installed. Now add the following code to .cshrc:\n"
+print "if ( -f $HOME/.cshrc.common ) then"
+print "    source $HOME/.cshrc.common"
+print "endif\n"
